+++ msgraph/models.go
@@ -44,6 +44,7 @@ type Application struct {
 	AddIns                        *[]AddIn                  `json:"addIns,omitempty"`
 	Api                           *ApplicationApi           `json:"api,omitempty"`
 	AppId                         *string                   `json:"appId,omitempty"`
+	ApplicationTemplateId         *string                   `json:"applicationTemplateId,omitempty"`
 	AppRoles                      *[]AppRole                `json:"appRoles,omitempty"`
 	CreatedDateTime               *time.Time                `json:"createdDateTime,omitempty"`
 	DefaultRedirectUri            *string                   `json:"defaultRedirectUri,omitempty"`
@@ -290,6 +291,21 @@ type ApplicationSpa struct {
 	RedirectUris *[]string `json:"redirectUris,omitempty"`
 }
 
+type ApplicationTemplate struct {
+	ID                         *string                        `json:"id,omitempty"`
+	Categories                 *[]ApplicationTemplateCategory `json:"categories,omitempty"`
+	Description                *string                        `json:"description,omitempty"`
+	DisplayName                *string                        `json:"displayName,omitempty"`
+	HomePageUrl                *string                        `json:"homePageUrl,omitempty"`
+	LogoUrl                    *string                        `json:"logoUrl,omitempty"`
+	Publisher                  *string                        `json:"publisher,omitempty"`
+	SupportedProvisioningTypes *[]string                      `json:"supportedProvisioningTypes,omitempty"`
+	SupportedSingleSignOnModes *[]string                      `json:"supportedSingleSignOnModes,omitempty"`
+
+	Application      *Application      `json:"application,omitempty"`
+	ServicePrincipal *ServicePrincipal `json:"servicePrincipal,omitempty"`
+}
+
 type ApplicationWeb struct {
 	HomePageUrl           *StringNullWhenEmpty   `json:"homePageUrl,omitempty"`
 	ImplicitGrantSettings *ImplicitGrantSettings `json:"implicitGrantSettings,omitempty"`
@@ -389,7 +405,7 @@ type ConditionalAccessPolicy struct {
 	ID               *string                           `json:"id,omitempty"`
 	ModifiedDateTime *time.Time                        `json:"modifiedDateTime,omitempty"`
 	SessionControls  *ConditionalAccessSessionControls `json:"sessionControls,omitempty"`
-	State            *string                           `json:"state,omitempty"`
+	State            *ConditionalAccessPolicyState     `json:"state,omitempty"`
 }
 
 type ConditionalAccessSessionControls struct {
@@ -707,14 +723,14 @@ type InformationalUrl struct {
 
 // Invitation describes a Invitation object.
 type Invitation struct {
-	ID                      *string `json:"id,omitempty"`
-	InvitedUserDisplayName  *string `json:"invitedUserDisplayName,omitempty"`
-	InvitedUserEmailAddress *string `json:"invitedUserEmailAddress,omitempty"`
-	SendInvitationMessage   *bool   `json:"sendInvitationMessage,omitempty"`
-	InviteRedirectURL       *string `json:"inviteRedirectUrl,omitempty"`
-	InviteRedeemURL         *string `json:"inviteRedeemUrl,omitempty"`
-	Status                  *string `json:"status,omitempty"`
-	InvitedUserType         *string `json:"invitedUserType,omitempty"`
+	ID                      *string          `json:"id,omitempty"`
+	InvitedUserDisplayName  *string          `json:"invitedUserDisplayName,omitempty"`
+	InvitedUserEmailAddress *string          `json:"invitedUserEmailAddress,omitempty"`
+	SendInvitationMessage   *bool            `json:"sendInvitationMessage,omitempty"`
+	InviteRedirectURL       *string          `json:"inviteRedirectUrl,omitempty"`
+	InviteRedeemURL         *string          `json:"inviteRedeemUrl,omitempty"`
+	Status                  *string          `json:"status,omitempty"`
+	InvitedUserType         *InvitedUserType `json:"invitedUserType,omitempty"`
 
 	InvitedUserMessageInfo *InvitedUserMessageInfo `json:"invitedUserMessageInfo,omitempty"`
 	InvitedUser            *User                   `json:"invitedUser,omitempty"`
@@ -947,7 +963,7 @@ func (se SchemaExtensionData) MarshalJSON() ([]byte, error) {
 // ServicePrincipal describes a Service Principal object.
 type ServicePrincipal struct {
 	DirectoryObject
-	Owners *Owners `json:"-"`
+	Owners *Owners `json:"owners@odata.bind,omitempty"`
 
 	AccountEnabled                      *bool                         `json:"accountEnabled,omitempty"`
 	AddIns                              *[]AddIn                      `json:"addIns,omitempty"`
@@ -1083,6 +1099,7 @@ type User struct {
 	DisplayName                     *string                  `json:"displayName,omitempty"`
 	EmployeeHireDate                *time.Time               `json:"employeeHireDate,omitempty"`
 	EmployeeId                      *StringNullWhenEmpty     `json:"employeeId,omitempty"`
+	EmployeeOrgData                 *EmployeeOrgData         `json:"employeeOrgData,omitempty"`
 	EmployeeType                    *string                  `json:"employeeType,omitempty"`
 	ExternalUserState               *string                  `json:"externalUserState,omitempty"`
 	FaxNumber                       *StringNullWhenEmpty     `json:"faxNumber,omitempty"`
@@ -1107,7 +1124,7 @@ type User struct {
 	OnPremisesSyncEnabled           *bool                    `json:"onPremisesSyncEnabled,omitempty"`
 	OnPremisesUserPrincipalName     *string                  `json:"onPremisesUserPrincipalName,omitempty"`
 	OtherMails                      *[]string                `json:"otherMails,omitempty"`
-	PasswordPolicies                *string                  `json:"passwordPolicies,omitempty"`
+	PasswordPolicies                *StringNullWhenEmpty     `json:"passwordPolicies,omitempty"`
 	PasswordProfile                 *UserPasswordProfile     `json:"passwordProfile,omitempty"`
 	PastProjects                    *[]string                `json:"pastProjects,omitempty"`
 	PostalCode                      *StringNullWhenEmpty     `json:"postalCode,omitempty"`
@@ -1238,4 +1255,9 @@ type WindowsHelloForBusinessAuthenticationMethod struct {
 	DisplayName     *string                          `json:"displayName,omitempty"`
 	ID              *string                          `json:"id,omitempty"`
 	KeyStrength     *AuthenticationMethodKeyStrength `json:"authenticationMethodKeyStrength,omitempty"`
+}
+
+type EmployeeOrgData struct {
+	CostCenter *string `json:"costCenter,omitempty"`
+	Division   *string `json:"division,omitempty"`
 }